Fix 'compressBound' argument type to 'toUInt()' in 'ZLIB.native.kt'

+++ kompress/native/src/me/devnatan/kompress/ZLIB.native.kt
@@ -57,7 +57,7 @@ internal actual fun zlibInflate(input: ByteArray): ByteArray =
 @OptIn(ExperimentalForeignApi::class)
 internal actual fun zlibDeflate(input: ByteArray): ByteArray =
     memScoped {
-        val compressedBuffer = ByteArray(compressBound(input.size.toULong()).toInt())
+        val compressedBuffer = ByteArray(compressBound(input.size.toUInt()).toInt())
 
         val stream = alloc<z_stream>()
         stream.avail_in = input.size.toUInt()
